plotcake: line_add_value(): if @limit < 0, no limit
Vers: Fedora Linux 44, Linux 7.0.12-201.fc44.x86_64, GCC 16.1.1, GLibc 2.43, Ncurses 6.6.20251230

+++ tools/ncurses/tools/value.h
@@ -35,7 +35,7 @@ struct line {
 	const char *name;
 	int color; /* C_RED, ... */
 	struct value *head, *tail, *max, *min;
-	int count; /* number of value */
+	long count; /* number of value */
 	struct line *next; /* maybe line in group */
 	struct lgroup *lg; /* belongs to */
 	const struct ldraw_ops *ops;
@@ -101,7 +101,7 @@ enum lcolor_enum color_name2n(const char *name);
 bool hascolor_name(const char *name);
 enum lcolor_enum nextcolor(enum lcolor_enum c);
 
-void line_add_value(struct line *l, double v, unsigned long limit);
+void line_add_value(struct line *l, double v, long limit);
 double line_range_max(struct line *l, int start, int len);
 double line_range_min(struct line *l, int start, int len);
